fix: only override on non darwin based systems

+++ archi.nix
@@ -3,16 +3,27 @@
 
 { pkgs }:
 
-with pkgs;
+if pkgs.stdenv.isDarwin then
+  pkgs.archi
+else
 
-writeShellApplication {
-  name = pkgs.archi.pname;
-  runtimeInputs = with pkgs; [ xvfb-run archi ];
+  pkgs.writeShellApplication {
+    name = pkgs.archi.pname;
+    runtimeInputs = with pkgs; [ xvfb-run archi ];
 
-  text = ''
-    Xvfb :99 &
-    export DISPLAY=:99
-    Archi "$@"
-    pkill -f 'Xvfb :99'
-  '';
-}
+    text = ''
+      Xvfb :99 &
+      export DISPLAY=:99
+      xvfb-run Archi -application com.archimatetool.commandline.app -consoleLog -nosplash "$@"
+      pkill -f 'Xvfb :99'
+    '';
+
+    meta = {
+      description = "xvfb wrapper around Archi";
+      longDescription = ''
+        This is a wrapper script that runs Archi in a headless environment using Xvfb.
+        It allows you to use Archi's command line interface on systems without a graphical display.
+      '';
+      mainProgram = pkgs.archi.name;
+    };
+  }
